API: Eliminate the whole concept of State
- Problem with State was described in https://github.com/raquo/Laminar/issues/37
- Simplify the hierarchy as well: merge LazyObservable into Observable and MemoryObservable into Signal
