+++ src/main/scala/com/raquo/airstream/signal/MapSignal.scala
@@ -1,7 +1,10 @@
 package com.raquo.airstream.signal
 
-import com.raquo.airstream.core.MemoryObservable
-import com.raquo.airstream.features.MapMemoryObservable
+import com.raquo.airstream.core.AirstreamError.ErrorHandlingError
+import com.raquo.airstream.core.Transaction
+import com.raquo.airstream.features.{InternalTryObserver, SingleParentObservable}
+
+import scala.util.{Failure, Success, Try}
 
 // @TODO[Elegance] Is this right that we shoved .recover into Map observables? Maybe it deserves its own class? But it's so many classes...
 
@@ -16,10 +19,62 @@ import com.raquo.airstream.features.MapMemoryObservable
   * @param recover Note: guarded against exceptions
   */
 class MapSignal[I, O](
-  override protected[this] val parent: MemoryObservable[I],
-  override protected[this] val project: I => O,
-  override protected[this] val recover: Option[PartialFunction[Throwable, Option[O]]]
-) extends Signal[O] with MapMemoryObservable[I, O] {
+  protected[this] val parent: Signal[I],
+  protected[this] val project: I => O,
+  protected[this] val recover: Option[PartialFunction[Throwable, Option[O]]]
+) extends Signal[O] with SingleParentObservable[I, O] with InternalTryObserver[I] {
 
   override protected[airstream] val topoRank: Int = parent.topoRank + 1
+
+  override protected[airstream] def onTry(nextParentValue: Try[I], transaction: Transaction): Unit = {
+    nextParentValue.fold(
+      nextError => recover.fold(
+        // if no `recover` specified, fire original error
+        fireError(nextError, transaction))(
+        pf => Try(pf.applyOrElse(nextError, (_: Throwable) => null)).fold(
+          tryError => {
+            // if recover throws error, fire wrapped error
+            fireError(ErrorHandlingError(error = tryError, cause = nextError), transaction)
+          },
+          nextValue => {
+            if (nextValue == null) {
+              // If recover was not applicable, fire original error
+              fireError(nextError, transaction)
+            } else {
+              // If recover was applicable and resulted in a new value, fire that value
+              nextValue.foreach(fireValue(_, transaction))
+            }
+          }
+        )
+      ),
+      _ => fireTry(nextParentValue.map(project), transaction)
+    )
+  }
+
+  override protected[this] def initialValue: Try[O] = {
+    val originalValue = parent.tryNow().map(project)
+
+    originalValue.fold(
+      // if no `recover` specified, keep original error
+      nextError => recover.fold(originalValue)(
+        pf => Try(pf.applyOrElse(nextError, (_: Throwable) => null)).fold(
+          tryError => {
+            // if recover throws error, save wrapped error
+            Failure(ErrorHandlingError(error = tryError, cause = nextError))
+          },
+          nextValue => {
+            if (nextValue == null) {
+              // If recover was not applicable, keep original value
+              originalValue
+            } else {
+              // @TODO[Test] Verify this
+              // If recover was applicable and resulted in Some(value), use that.
+              // If None, use the original error because we can't "skip" initial value
+              nextValue.map(Success(_)).getOrElse(originalValue)
+            }
+          }
+        )
+      ),
+      _ => originalValue)
+  }
 }
