+++ src/main/scala/com/raquo/airstream/signal/SignalViewer.scala
@@ -7,8 +7,6 @@ import scala.util.Try
 
 /** This class adds a noop observer to `signal`, ensuring that its current value is computed.
   * It then lets you query `signal`'s current value with `now` and `tryNow` methods.
-  *
-  * This will likely replace the entirety of the `State` type in Airstream.
   */
 class SignalViewer[+A](signal: Signal[A], owner: Owner) {
