+++ src/test/scala/com/raquo/airstream/eventstream/SampleCombineEventStream2Spec.scala
@@ -136,124 +136,4 @@ class SampleCombineEventStream2Spec extends FunSpec with Matchers {
     calculations.clear()
     effects.clear()
   }
-
-  it("gets current value of State") {
-
-    implicit val testOwner: TestableOwner = new TestableOwner
-
-    val bus1 = new EventBus[Int]
-    val bus2 = new EventBus[Int]
-
-    val calculations = mutable.Buffer[Calculation[Int]]()
-    val effects = mutable.Buffer[Effect[Int]]()
-
-    val sampledState = bus2.events.toState(0).map(Calculation.log("state", calculations))
-
-    val combinedStream = bus1.events
-      .map(Calculation.log("stream", calculations))
-      .withCurrentValueOf(sampledState)
-      .map2(_ + _)
-      .map(Calculation.log("combined", calculations))
-
-    val combinedObserver = Observer[Int](effects += Effect("combined", _))
-
-    // --
-
-    calculations shouldEqual mutable.Buffer(
-      Calculation("state", 0) // eager initial value
-    )
-    effects shouldEqual mutable.Buffer()
-
-    calculations.clear()
-
-    // --
-
-    val subCombined = combinedStream.addObserver(combinedObserver)
-
-    calculations shouldEqual mutable.Buffer()
-    effects shouldEqual mutable.Buffer()
-
-    calculations.clear()
-
-    // --
-
-    bus1.writer.onNext(1)
-
-    calculations shouldEqual mutable.Buffer(
-      Calculation("stream", 1),
-      Calculation("combined", 1)
-    )
-    effects shouldEqual mutable.Buffer(
-      Effect("combined", 1)
-    )
-
-    calculations.clear()
-    effects.clear()
-
-    // --
-
-    bus2.writer.onNext(100)
-
-    calculations shouldEqual mutable.Buffer(
-      Calculation("state", 100)
-    )
-    effects shouldEqual mutable.Buffer()
-
-    calculations.clear()
-
-    // --
-
-    bus1.writer.onNext(2)
-
-    calculations shouldEqual mutable.Buffer(
-      Calculation("stream", 2),
-      Calculation("combined", 102)
-    )
-    effects shouldEqual mutable.Buffer(
-      Effect("combined", 102)
-    )
-
-    calculations.clear()
-    effects.clear()
-
-    // --
-
-    subCombined.kill()
-
-    calculations shouldEqual mutable.Buffer()
-    effects shouldEqual mutable.Buffer()
-
-    // --
-
-    bus2.writer.onNext(200)
-
-    calculations shouldEqual mutable.Buffer(
-      Calculation("state", 200)
-    )
-    effects shouldEqual mutable.Buffer()
-
-    calculations.clear()
-
-    // --
-
-    combinedStream.addObserver(combinedObserver)
-
-    calculations shouldEqual mutable.Buffer()
-    effects shouldEqual mutable.Buffer()
-
-    // --
-
-    bus1.writer.onNext(3)
-
-    calculations shouldEqual mutable.Buffer(
-      Calculation("stream", 3),
-      Calculation("combined", 203)
-    )
-    effects shouldEqual mutable.Buffer(
-      Effect("combined", 203)
-    )
-
-    calculations.clear()
-    effects.clear()
-  }
 }
